double är bättre än float

+++ spring-docker-demo/src/main/java/com/example/springdocker/model/Drink.java
@@ -20,7 +20,7 @@ public class Drink {
         public int multiply(int a, int b){
                 return a*b;
         }
-        public float divide(float a, float b){
-                return a/b;
+        public double divide(double a, double b){
+          return a/b;
         }
 }
